requirements.txt
# Backend requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
chromadb==0.4.15
python-multipart==0.0.6
asyncio==3.4.3
langchain==0.0.285
python-dotenv==1.0.0
httpx==0.24.1
uuid==1.30
anthropic==0.5.0
openai==0.28.1
tiktoken==0.5.1
pandas==2.1.1
numpy==1.26.0
pdf2image==1.16.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==9.5.0
sqlalchemy==2.0.21
aiofiles==23.1.0
# Frontend requirements.txt
streamlit==1.27.0
pandas==2.1.1
numpy==1.26.0
plotly==5.17.0
streamlit-extras==0.3.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
httpx==0.24.1
pydantic==2.3.0
python-dotenv==1.0.0